# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hariri Foundation API",
        "description": "API documentation for Users Management",
        "contact": {
            "email": "support@haririfoundation.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/responses": {
            "get": {
                "tags": [
                    "Responses"
                ],
                "summary": "Get all responses or a specific response by ID",
                "operationId": "d8ddb7326e4cb824e09fa87e2e2b6d10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Optional response UUID to fetch a specific record",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Responses retrieved successfully"
                    },
                    "404": {
                        "description": "Response not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Responses"
                ],
                "summary": "Create a new response record",
                "operationId": "b30acd802a2f15906cac094ae4fc39d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "survey_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "survey_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "ext_12345"
                                    },
                                    "submitted_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response created successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/responses/{id}": {
            "put": {
                "tags": [
                    "Responses"
                ],
                "summary": "Update an existing response",
                "operationId": "d85a2400881b700ca9c7d4e367c38b42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Response UUID to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "survey_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "ext_12345"
                                    },
                                    "submitted_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response updated successfully"
                    },
                    "404": {
                        "description": "Response not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Responses"
                ],
                "summary": "Delete a response by UUID",
                "operationId": "412d98d4f02839192a33ee19ef05c345",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Response UUID to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response deleted successfully"
                    },
                    "404": {
                        "description": "Response not found"
                    }
                }
            }
        },
...
    },
...
}